Add a `!highlight` command to track stream highlights
Streamers want a way of recording the greatest moments of their stream
without having to collect all the various clips folks make; this commit
marks the start of a !highlight command, modeled on LRRBot's, to do so.

`!highlight foo` currently makes the bot output a link to the vod, the
current uptime of the stream, the user logging it, the context given
('foo' in this case) and (as of right now) write it back to the channel.
In the future, this needs to write to a datastore of some kind (Google
Sheets seems the preferred method) to allow for collaborative CRUD
operations.